+++ src/main/java/rs/raf/repositories/comment/CommentRepository.java
@@ -7,6 +7,6 @@ import java.util.List;
 public interface CommentRepository {
 
     public Comment addComment(Comment comment);
-    public Comment findComment(Integer id);
+    public Comment findComment(Long id);
     public List<Comment> allComments();
 }
